wembleyCalendar/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
ratelimit==2.2.1
requests==2.32.3
setuptools==69.5.1
soupsieve==2.5
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
